till Q 7

+++ PE2/src/com/stackroute/pe2/practiseQ2.java
@@ -6,7 +6,7 @@ import java.lang.Math;
 public class practiseQ2 {
     public static void main(String srgs[]){
         Scanner sc=new Scanner(System.in);
-        int var=sc.nextInt();
+        long var=sc.nextLong();
         practiseQ2 obj=new  practiseQ2();
 
         if(obj.mylog(var,4)==(int)obj.mylog(var,4)){
@@ -16,7 +16,7 @@ public class practiseQ2 {
             System.out.println("Number is not power of 4");
     }
 
-    public double mylog(int a,int b){
+    public double mylog(long a,int b){
         return(Math.log(a)/Math.log(b));
     }
 }
